backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.1
defusedxml==0.7.1
Django
django-autoslug==1.9.8
django-environ==0.8.1
django-filter
django-slugger==1.1.2
django-templated-mail==1.1.1
django-uuslug==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow
pycparser==2.21
PyJWT==2.3.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
webcolors==1.11.1
djangorestframework_simplejwt
reportlab
gunicorn
python-dotenv==0.19.2